Adopt class based routes

+++ src/routes/web.php
@@ -1,12 +1,15 @@
 <?php
 
+use App\Http\Controllers\Admin;
+use App\Http\Controllers\Home;
+use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Config;
 
 /***
  * CKEditor
  */
-Route::get('elfinder/ckeditor', '\Barryvdh\Elfinder\ElfinderController@showCKeditor4');
-Route::get('elfinder/popup', '\Barryvdh\Elfinder\Barryvdh\Elfinder\ElfinderController@showPopup');
+Route::get('elfinder/ckeditor', [\Barryvdh\Elfinder\ElfinderController::class, 'showCKeditor4']);
+Route::get('elfinder/popup', [\Barryvdh\Elfinder\Barryvdh\Elfinder\ElfinderController::class, 'showPopup']);
 
 /***
  * Authen page
@@ -18,22 +21,22 @@ Route::prefix(LaravelLocalization::setLocale())->middleware('web')->group(functi
 /***
  * Admin page
  */
-Route::prefix(LaravelLocalization::setLocale().'/admin')->middleware('web', 'auth', 'localeSessionRedirect', 'localizationRedirect')->namespace('Admin')->group(
+Route::prefix(LaravelLocalization::setLocale().'/admin')->middleware('web', 'auth', 'localeSessionRedirect', 'localizationRedirect')->group(
     function () {
-        Route::get('profile', 'ProfileController@index');
-        Route::post('profile', 'ProfileController@update');
+        Route::get('profile', [Admin\ProfileController::class, 'index']);
+        Route::post('profile', [Admin\ProfileController::class, 'update']);
 
-        Route::get('/', 'WelcomeController@index');
-        Route::get('settings', 'SettingController@index');
-        Route::post('settings', 'SettingController@update');
+        Route::get('/', [Admin\WelcomeController::class, 'index']);
+        Route::get('settings', [Admin\SettingController::class, 'index']);
+        Route::post('settings', [Admin\SettingController::class, 'update']);
 
-        Route::get('media', '\Barryvdh\Elfinder\ElfinderController@showCKeditor4');
+        Route::get('media', [\Barryvdh\Elfinder\ElfinderController::class, 'showCKeditor4']);
 
-        Route::get('banner', 'BannerController@index');
-        Route::post('banner', 'BannerController@update');
+        Route::get('banner', [Admin\BannerController::class, 'index']);
+        Route::post('banner', [Admin\BannerController::class, 'update']);
 
-        Route::get('issues', 'IssueController@index');
-        Route::delete('issues/{id}', 'IssueController@destroy');
+        Route::get('issues', [Admin\IssueController::class, 'index']);
+        Route::delete('issues/{id}', [Admin\IssueController::class, 'destroy']);
 
         // GET DYNAMIC ROUTE =======================================================
         $listsMenu = Config::get('setting.lists');
@@ -42,24 +45,24 @@ Route::prefix(LaravelLocalization::setLocale().'/admin')->middleware('web', 'aut
         // ROUTE CATEGORY MENU
         foreach ($catcontentsMenu as $key => $content) {
             $routecat = $key.'/categories';
-            Route::get($routecat, 'CategoryController@index');
-            Route::resource($routecat, 'CategoryController');
+            Route::get($routecat, [Admin\CategoryController::class, 'index']);
+            Route::resource($routecat, Admin\CategoryController::class);
 
             $routeproduct = $key.'/product';
-            Route::get($routeproduct, 'ProductController@index');
-            Route::resource($routeproduct, 'ProductController');
+            Route::get($routeproduct, [Admin\ProductController::class, 'index']);
+            Route::resource($routeproduct, Admin\ProductController::class);
         }
 
         // ROUTE CONTENTS MENU
-        Route::get('content/{menu}', 'ContentController@index');
-        Route::post('content/{menu}', 'ContentController@update');
+        Route::get('content/{menu}', [Admin\ContentController::class, 'index']);
+        Route::post('content/{menu}', [Admin\ContentController::class, 'update']);
 
         // ROUTE LISTS MENU
         foreach ($listsMenu as $key => $content) {
             $routesub = 'lists/'.$key;
 
-            Route::get($routesub, 'ListController@index');
-            Route::resource($routesub, 'ListController');
+            Route::get($routesub, [Admin\ListController::class, 'index']);
+            Route::resource($routesub, Admin\ListController::class);
         }
         // =========================================================================
 
@@ -88,53 +91,53 @@ Route::prefix(LaravelLocalization::setLocale().'/admin')->middleware('web', 'aut
         // Route::resource('download/gallerys', 'GallerysController');
 
         // MANAGE STUDENT
-        Route::get('student', 'StudentController@index');
-        Route::post('student/save', 'StudentController@save');
-        Route::resource('student', 'StudentController');
+        Route::get('student', [Admin\StudentController::class, 'index']);
+        Route::post('student/save', [Admin\StudentController::class, 'save']);
+        Route::resource('student', Admin\StudentController::class);
 
         // PAYMENT
-        Route::get('payment', 'PaymentMastersController@index');
-        Route::resource('payment', 'PaymentMastersController');
-        Route::post('payment/save', 'PaymentMastersController@save');
-        Route::post('payment/batchdelete', 'PaymentMastersController@batchdelete');
-        Route::get('paymentdetail', 'PaymentDetailsController@index');
-        Route::resource('paymentdetail', 'PaymentDetailsController');
-        Route::post('paymentdetail/save', 'PaymentDetailsController@save');
+        Route::get('payment', [Admin\PaymentMastersController::class, 'index']);
+        Route::resource('payment', Admin\PaymentMastersController::class);
+        Route::post('payment/save', [Admin\PaymentMastersController::class, 'save']);
+        Route::post('payment/batchdelete', [Admin\PaymentMastersController::class, 'batchdelete']);
+        Route::get('paymentdetail', [Admin\PaymentDetailsController::class, 'index']);
+        Route::resource('paymentdetail', Admin\PaymentDetailsController::class);
+        Route::post('paymentdetail/save', [Admin\PaymentDetailsController::class, 'save']);
 
         // IMPORT
         // Route::get('import/custInfo', 'ImportController@custinfo');
         // Route::get('import/invoice', 'ImportController@invoice');
-        Route::post('import/uploadcustinfo', 'ImportController@uploadCustInfo');
-        Route::post('import/uploadinvoice', 'ImportController@uploadInvoice');
+        Route::post('import/uploadcustinfo', [Admin\ImportController::class, 'uploadCustInfo']);
+        Route::post('import/uploadinvoice', [Admin\ImportController::class, 'uploadInvoice']);
 
         // Mail
-        Route::get('template', 'TemplatesController@index');
-        Route::resource('template', 'TemplatesController');
-        Route::get('maillist', 'MaillistsController@index');
-        Route::resource('maillist', 'MaillistsController');
+        Route::get('template', [Admin\TemplatesController::class, 'index']);
+        Route::resource('template', Admin\TemplatesController::class);
+        Route::get('maillist', [Admin\MaillistsController::class, 'index']);
+        Route::resource('maillist', Admin\MaillistsController::class);
 
         // Report
-        Route::get('report', 'ReportsController@index');
+        Route::get('report', [Admin\ReportsController::class, 'index']);
         // Route::resource('report', 'ReportsController');
-        Route::get('report/export', 'ReportsController@export');
+        Route::get('report/export', [Admin\ReportsController::class, 'export']);
 
         // gen mail
-        Route::post('maillist/genmail', 'MaillistsController@genmail');
-        Route::post('maillist/sendmail', 'MaillistsController@sendmail');
-        Route::post('maillist/getfile', 'MaillistsController@getfile');
+        Route::post('maillist/genmail', [Admin\MaillistsController::class, 'genmail']);
+        Route::post('maillist/sendmail', [Admin\MaillistsController::class, 'sendmail']);
+        Route::post('maillist/getfile', [Admin\MaillistsController::class, 'getfile']);
 
         // users
-        Route::get('users', 'UserController@index');
-        Route::resource('users', 'UserController');
+        Route::get('users', [Admin\UserController::class, 'index']);
+        Route::resource('users', Admin\UserController::class);
     });
 
 /***
  * Home page
  */
-Route::prefix(LaravelLocalization::setLocale())->middleware('web', 'localeSessionRedirect', 'localizationRedirect')->namespace('Home')->group(
+Route::prefix(LaravelLocalization::setLocale())->middleware('web', 'localeSessionRedirect', 'localizationRedirect')->group(
     function () {
-        Route::get('/', 'HomeController@index');
-        Route::get('home', 'HomeController@index');
+        Route::get('/', [Home\HomeController::class, 'index']);
+        Route::get('home', [Home\HomeController::class, 'index']);
 
         // CONTACT
         // Route::get('contact', 'ContactController@index');
@@ -144,22 +147,22 @@ Route::prefix(LaravelLocalization::setLocale())->middleware('web', 'localeSessio
         //  return $captcha->src($config);
         // });
 
-        Route::get('invoice/{studentid}/{masterid}', 'HomeController@invoice');
-        Route::get('payment/{encryptionstr}', 'HomeController@payment');
+        Route::get('invoice/{studentid}/{masterid}', [Home\HomeController::class, 'invoice']);
+        Route::get('payment/{encryptionstr}', [Home\HomeController::class, 'payment']);
 
-        Route::get('payment/result/success', 'HomeController@paymentsuccess');
-        Route::get('payment/result/fail', 'HomeController@paymentfail');
-        Route::get('payment/result/cancel', 'HomeController@paymentcancel');
+        Route::get('payment/result/success', [Home\HomeController::class, 'paymentsuccess']);
+        Route::get('payment/result/fail', [Home\HomeController::class, 'paymentfail']);
+        Route::get('payment/result/cancel', [Home\HomeController::class, 'paymentcancel']);
         // Route::post('confirm', 'HomeController@paymentconfirm');
         // Route::get('paymentinfo', 'HomeController@paymentinfo');
     });
 
-Route::middleware('web')->namespace('Home')->group(function () {
-    Route::get('/set/{value}', 'SessionController@set');
-    Route::get('/get', 'SessionController@get');
-    Route::get('/destroy', 'SessionController@destroy');
+Route::middleware('web')->group(function () {
+    Route::get('/set/{value}', [Home\SessionController::class, 'set']);
+    Route::get('/get', [Home\SessionController::class, 'get']);
+    Route::get('/destroy', [Home\SessionController::class, 'destroy']);
 
-    Route::get('/en/set/{value}', 'SessionController@set');
-    Route::get('/en/get', 'SessionController@get');
-    Route::get('/en/destroy', 'SessionController@destroy');
+    Route::get('/en/set/{value}', [Home\SessionController::class, 'set']);
+    Route::get('/en/get', [Home\SessionController::class, 'get']);
+    Route::get('/en/destroy', [Home\SessionController::class, 'destroy']);
 });
